feat(UI): improvements #3
