+++ app/src/components/Layout/components/Sidebar/types.ts
@@ -1,0 +1,5 @@
+export type SidebarProps = {
+  width: number;
+  onResize: () => void;
+  resized: boolean;
+};
