+++ app/src/components/Layout/components/Topbar/index.ts
@@ -1,0 +1,3 @@
+import Topbar from './Topbar';
+
+export default Topbar;
